requirements.txt
torch==1.10.0+cu111

torchaudio==0.10.0+cu111
librosa==0.8.1

pandas==1.3.4
numpy==1.20.3
scipy==1.7.3

googledrivedownloader==0.4
matplotlib==3.5.0
wandb==0.12.7
tqdm==4.62.3
Pillow==8.4.0